Refactor: multiple 오류 문제 해결
