+++ frontend/src/components/Layout/AppLayout.tsx
@@ -1,5 +1,5 @@
-import Footer from "../Footer/Footer";
-import Header from "../Header/Header";
+import Footer from "./Footer";
+import Header from "./Header";
 import styles from "./AppLayout.module.css";
 
 const AppLayout = (props: { children: React.ReactNode }) => {
