+++ frontend/src/components/Layout/Header.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import React from "react";
-import styles from "./header.module.css";
+import styles from "./Header.module.css";
 import { BsSearch } from "react-icons/bs";
 import { FaUserCircle } from "react-icons/fa";
 import styled from "styled-components";
